decoupled App from index file
