+++ client/src/components/App.jsx
@@ -1,0 +1,51 @@
+import React, { Component } from 'react';
+import axios from 'axios';
+
+import ReviewsSummary from './ReviewsSummary.jsx';
+import TopReviews from './TopReviews.jsx';
+import ReviewsFeed from './ReviewsFeed.jsx';
+
+class App extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            productID: 21,
+            productReviews: [],
+        };
+
+    };
+
+    getPositiveReviews(reviews) {
+        return reviews.filter(review => review.rating > 3)
+    };
+
+    getCriticalReviews(reviews) {
+        return reviews.filter(review => review.rating < 3)
+    };
+
+    getReviews() {
+        axios.get(`/reviews/${this.state.productID}`)
+        .then((response) => {
+            this.setState({ productReviews: response.data })
+        })
+        .catch((err) => console.log(err))
+    };
+
+    componentDidMount() {
+        this.getReviews();
+    };
+
+    render() {
+        const positiveReviews = this.getPositiveReviews(this.state.productReviews);
+        const criticalReviews = this.getCriticalReviews(this.state.productReviews);
+        return (
+            <div>
+                < ReviewsSummary reviews={this.state.productReviews} />
+                < TopReviews positiveReviews={positiveReviews} criticalReviews={criticalReviews} />
+                < ReviewsFeed reviews={this.state.productReviews} />
+            </div>
+        )
+    };
+};
+
+export default App;
